rename temp component, restructure components, refactor hamburger menu
