+++ src/components/display/PageInConstruction/PageInConstruction.tsx
@@ -1,0 +1,14 @@
+import { Snackbar } from '@mui/base';
+import { Alert } from '@mui/material';
+import * as React from 'react';
+
+function PageInConstruction() {
+	return (
+		<Snackbar open={true}>
+			<Alert severity='warning' sx={{ width: '100%' }}>
+				{`This page is under construction, mind your head...`}
+			</Alert>
+		</Snackbar>
+	);
+}
+export { PageInConstruction };
