+++ src/components/display/SwapText/SwapText.tsx
