+++ src/constants/about-constants.ts
@@ -1,0 +1,19 @@
+const podcasts = [
+	{
+		text: 'Fall of Civilizations',
+		link: 'https://fallofcivilizationspodcast.com/',
+		img: 'fall_of_civs.jpg',
+	},
+	{
+		text: 'History of Rome',
+		link: 'https://thehistoryofrome.typepad.com/',
+		img: 'The_History_of_Rome.png',
+	},
+	{
+		text: 'The Rest is History',
+		link: 'https://therestishistory.com/episodes/',
+		img: 'rest_is_hist.jpg',
+	},
+];
+
+export { podcasts };
