--- src/constants/constants.ts
@@ -1,42 +1,0 @@
-import GitHubIcon from '@mui/icons-material/GitHub';
-import LinkedInIcon from '@mui/icons-material/LinkedIn';
-import { SvgIconComponent } from '@mui/icons-material';
-
-interface ExternalLink {
-	href: string;
-	label: string;
-	icon: SvgIconComponent;
-}
-
-const externalLinks: ExternalLink[] = [
-	{
-		href: 'https://www.linkedin.com/in/david-mcalister/',
-		label: 'LinkedIn',
-		icon: LinkedInIcon,
-	},
-	{
-		href: 'https://github.com/dalapto',
-		label: 'GitHub',
-		icon: GitHubIcon,
-	},
-];
-
-const podcasts = [
-	{
-		text: 'Fall of Civilizations',
-		link: 'https://fallofcivilizationspodcast.com/',
-		img: 'fall_of_civs.jpg',
-	},
-	{
-		text: 'History of Rome',
-		link: 'https://thehistoryofrome.typepad.com/',
-		img: 'The_History_of_Rome.png',
-	},
-	{
-		text: 'The Rest is History',
-		link: 'https://therestishistory.com/episodes/',
-		img: 'rest_is_hist.jpg',
-	},
-];
-
-export { externalLinks, podcasts };
