+++ src/constants/home-constants.ts
@@ -1,0 +1,44 @@
+const welcomes = [
+	'welcome',
+	'fáilte',
+	'croeso',
+	'bienvenue',
+	'willkommen',
+	'bienvenido',
+	'ようこそ',
+	'welkom',
+	'chào mừng',
+	'benvenuta',
+	'欢迎',
+	'mirë se erdhe',
+	'tervetuloa',
+	'welina',
+	'خوش آمدید',
+	'välkommen',
+	'добро пожаловать',
+	'olandiridwa',
+	'환영',
+	'Прошу',
+	'mile widziany',
+	'tunngasugit',
+	'vitaj',
+	'ברוך הבא',
+	'selamat datang',
+	'nabata',
+	'laipni lūdza',
+	'स्वागत है',
+	'maraba',
+	'hoş geldin',
+	'مەرھابا',
+	'byenvini',
+	'merħba',
+	`tun'ngahuktitauyut`,
+	'vælkomin',
+	'kαλώς ήρθες',
+	'أهلًا وسهلًا',
+	'benvingut',
+	'wamkelekile',
+	'歡迎',
+];
+
+export { welcomes };
