+++ src/constants/translations.ts
@@ -1,1 +1,1 @@
-//somewhere for text to live
+//somewhere for text to live ?
