+++ src/pages/Projects/RON/GrandCampaign.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Temp } from '../../../components/common/Temp/Temp';
+import { PageInConstruction } from '../../../components/display/PageInConstruction/PageInConstruction';
 
 function GrandCampaign() {
 	return (
 		<div className="App">
 			{' '}
-			<Temp></Temp>
+			<PageInConstruction></PageInConstruction>
 		</div>
 	);
 }
