+++ src/pages/Projects/m2tw/m2ex.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Temp } from '../../../components/common/Temp/Temp';
+import { PageInConstruction } from '../../../components/display/PageInConstruction/PageInConstruction';
 
 function m2ex() {
 	return (
 		<div className="App">
-			<Temp></Temp>
+			<PageInConstruction></PageInConstruction>
 		</div>
 	);
 }
